fixed basically unlimited cache folder depth
in exchange for basically unlimited folders in the first level of the
cache directory. The old cache directory contained too many nested levels
or folders for e.g. docker to handle which could lead to docker daemon
hanging and needed a complete restart. The levels of folders depended on
the length of the source image url and the number and values of options
for actual conversion. From now on the cache directory can contain sha256
number of folders with meta/source file and a "targets" folder with as
many files as there are combinations of (sorted) options to generate
thumbs etc. (depending on source staleness, caching and garbage
collection settings). New folder structure is basically this:

```
cache/
├── 3d38cda2f60d38c45b719c71128fc7d438a19e828d7cf4ac5f8c65d52a8be95c
│   ├── meta
│   ├── source
│   └── targets
│       ├── b24a174707f0eb7c6fa6ade7db303a767dfc6370af277f9f351c2cc0c0282b7a
│       └── __no_options
└── c0ee1e228a5e35561d39774005aa581eabe78903a810e84ecb4c56892c378ba3
    ├── meta
    ├── source
    └── targets
        ├── b24a174707f0eb7c6fa6ade7db303a767dfc6370af277f9f351c2cc0c0282b7a
        ├── dc95f52f1b3c5b20e20dab0afc79d981a9aaf7141f504939b76fafb2ed442852
        └── fd75cb709ec1c0b4ab756dd231ba945a950032b63bbb9a88c2fab32e2ee606ad
```

refs #138