bump deps, handle DEP0005, make npm audit happy
The "new Buffer(…)" usages have been fixed to non-deprecated
"Buffer.from(…)". The npm audit report container a vulnerabiility
for a lib that was used by nodeunit – that's why a specific sshpk
version is being mentioned in the devDependencies and nodeunit was moved
to that section as well. Updated changelog and shrinkwrap file and
allowed more engine versions in package file as tests run on all node
versions from 4.x up to current edge version of 12.x. refs #138